requirements.txt
fastapi
setuptools
wheel
uvicorn
langchain
langchain_community
langchain-anthropic
spacy
pydantic
python-dotenv